Hello World
This is my complete setup for the Ender 5 Plus on an BTT SKR 3 using the stock E5+ LCD screen, using the Desuuuu dwin LCD drivers. No fine tuning has been done, and likely has many mistakes on the setup for this machine. I'll work on it and update as I work out how to make things play nice :)